backend/common/requirements.txt
Flask==3.0.1
pymongo==3.11
pyyaml
requests